refactor: table 패키지 분리
