+++ src/main/java/kitchenpos/table/domain/JpaTableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.table.domain;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaTableGroupRepository extends JpaRepository<TableGroup, Long> {
