+++ src/main/java/kitchenpos/table/ui/request/OrderTableEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.request;
+package kitchenpos.table.ui.request;
 
 public class OrderTableEmptyRequest {
     private boolean empty;
