+++ src/main/java/kitchenpos/table/ui/request/OrderTableNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.request;
+package kitchenpos.table.ui.request;
 
 public class OrderTableNumberOfGuestsRequest {
     private int numberOfGuests;
