+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -12,20 +12,20 @@ import java.util.Optional;
 import kitchenpos.dao.JpaMenuGroupRepository;
 import kitchenpos.menu.domain.JpaMenuRepository;
 import kitchenpos.dao.JpaOrderRepository;
-import kitchenpos.dao.JpaOrderTableRepository;
+import kitchenpos.table.domain.JpaOrderTableRepository;
 import kitchenpos.product.domain.JpaProductRepository;
-import kitchenpos.dao.JpaTableGroupRepository;
+import kitchenpos.table.domain.JpaTableGroupRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.product.domain.Product;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 import kitchenpos.ui.dto.request.OrderLineItemRequest;
 import kitchenpos.ui.dto.request.OrderRequest;
-import kitchenpos.ui.dto.request.OrderTableIdRequest;
+import kitchenpos.table.ui.request.OrderTableIdRequest;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
