+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -3,6 +3,7 @@ package kitchenpos.application;
 import static kitchenpos.application.exception.ExceptionType.INVALID_PROCEEDING_TABLE_GROUP_EXCEPTION;
 import static kitchenpos.application.exception.ExceptionType.NOT_FOUND_TABLE_EXCEPTION;
 
+import kitchenpos.table.application.TableService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
