+++ src/test/java/kitchenpos/ui/RestControllerTest.java
@@ -6,9 +6,11 @@ import kitchenpos.menu.application.MenuService;
 import kitchenpos.application.OrderService;
 import kitchenpos.menu.ui.MenuRestController;
 import kitchenpos.product.application.ProductService;
-import kitchenpos.application.TableGroupService;
-import kitchenpos.application.TableService;
+import kitchenpos.table.application.TableGroupService;
+import kitchenpos.table.application.TableService;
 import kitchenpos.product.ui.ProductRestController;
+import kitchenpos.table.ui.TableGroupRestController;
+import kitchenpos.table.ui.TableRestController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
